+++ app/helpers/view_helper.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module ViewHelper
-  def money_presenter(cents, currency)
-    "#{Money::Currency.new(currency).symbol} #{Money.new(cents).round}"
+  def money_presenter(price, currency)
+    "#{Money::Currency.new(currency).symbol} #{price}"
   end
 end