+++ app/models/users/position.rb
@@ -7,8 +7,6 @@ module Users
     belongs_to :portfolio
     belongs_to :quote
 
-    monetize :price_cents
-
     scope :buying, -> { where selling_position: false }
     scope :selling, -> { where selling_position: true }
     scope :with_unsold_securities, -> { where sold_all: false }