+++ app/services/analytics/balance_service.rb
@@ -17,8 +17,8 @@ module Analytics
     def perform_analytics
       Portfolios::Cash.balance.where(portfolio: @portfolio_ids).each do |cash|
         currency_symbol = cash.amount_currency.to_sym
-        @result[:stats][currency_symbol] += cash.amount_cents
-        @result[:summary_cents] += cash.amount_cents * @exchange_rates[currency_symbol]
+        @result[:stats][currency_symbol] += cash.amount_cents / 100
+        @result[:summary_price] += cash.amount_cents * @exchange_rates[currency_symbol] / 100
       end
     end
 
@@ -29,7 +29,7 @@ module Analytics
           USD: 0,
           EUR: 0
         },
-        summary_cents: 0
+        summary_price: 0
       }
     end
   end