+++ app/services/analytics/share_sectors_service.rb
@@ -18,7 +18,7 @@ module Analytics
 
     def initialize_variables
       @result = {}
-      @total_amount_cents = 0
+      @total_amount_price = 0
     end
 
     def collect_absolute_values(stats)
@@ -26,19 +26,19 @@ module Analytics
         sector = quote.security.sector
         next unless sector
 
-        cents_amount = cents_amount(quote, stats)
-        sector_name  = sector_name(sector)
+        amount = amount(quote, stats)
+        sector_name = sector_name(sector)
 
-        update_total_cents_counter(cents_amount)
-        next update_sector(sector_name, cents_amount) if @result.has_key?(sector_name)
+        update_total_price_counter(amount)
+        next update_sector(sector_name, amount) if @result.has_key?(sector_name)
 
-        add_sector(sector_name, sector.color, cents_amount)
+        add_sector(sector_name, sector.color, amount)
       end
     end
 
     def count_relative_values
       @result.each do |_key, value|
-        value[:amount] = (100.0 * value[:amount] / @total_amount_cents).round(2)
+        value[:amount] = (100.0 * value[:amount] / @total_amount_price).round(2)
       end
     end
 
@@ -46,24 +46,24 @@ module Analytics
       @result = @result.sort_by { |_key, value| -value[:amount] }.to_h
     end
 
-    def update_total_cents_counter(cents_amount)
-      @total_amount_cents += cents_amount
+    def update_total_price_counter(amount)
+      @total_amount_price += amount
     end
 
-    def add_sector(sector_name, sector_color, cents_amount)
-      @result[sector_name] = { color: sector_color, amount: cents_amount }
+    def add_sector(sector_name, sector_color, amount)
+      @result[sector_name] = { color: sector_color, amount: amount }
     end
 
-    def update_sector(sector_name, cents_amount)
-      @result[sector_name][:amount] += cents_amount
+    def update_sector(sector_name, amount)
+      @result[sector_name][:amount] += amount
     end
 
     def sector_name(sector)
       sector.name[I18n.locale.to_s]
     end
 
-    def cents_amount(quote, stats)
-      stats[:selling_total_cents] * @exchange_rates[quote.price_currency.to_sym]
+    def amount(quote, stats)
+      stats[:selling_total_price] * @exchange_rates[quote.price_currency.to_sym]
     end
   end
 end