+++ spec/models/users/position_spec.rb
@@ -11,6 +11,4 @@ describe Users::Position, type: :model do
     it { is_expected.to belong_to :portfolio }
     it { is_expected.to belong_to :quote }
   end
-
-  it { is_expected.to monetize(:price) }
 end